sentigraph_backend/requirements.txt
Django==5.2
psycopg2-binary==2.9.9
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
channels==4.0.0
daphne==4.0.0
pandas==2.2.3
kagglehub==0.3.12

transformers[sentencepiece]==4.51.3
torch==2.7.0
huggingface_hub[hf_xet]==0.30.2